requirements.txt
Django==4.0
django_environ==0.9.0
django_filter==21.1
djangorestframework==3.13.1
environ==1.0
